+++ package.json
@@ -30,7 +30,6 @@
     "buffer": "^6.0.3",
     "core-js": "^3.6.5",
     "crypto-browserify": "^3.12.0",
-    "date-fns": "^2.27.0",
     "ethers": "^5.4.7",
     "naive-ui": "^2.19.11",
     "postcss": "^7",
